+++ src/openlayers/mapping/ImageSuperMapRest.js
@@ -19,7 +19,7 @@ import {Util} from '../core/Util';
  *        serverType - {SuperMap.ServerType} 服务类型。<br>
  *        redirect - {boolean} 是否重定向，默认为false。<br>
  *        transparent - {boolean} 图片是否透明，默认为true。<br>
- *        _cache - {boolean} 是否使用缓存，默认为true。<br>
+ *        cacheEnabled - {boolean} 是否使用服务端的缓存，默认为 true，即使用服务端的缓存。<br>
  *        prjCoordSys - {Object} 请求的地图的坐标参考系统。当此参数设置的坐标系统不同于地图的原有坐标系统时， 系统会进行动态投影，并返回动态投影后的地图图片。例如：{"epsgCode":3857}。<br>
  *        layersID - {string} 获取进行切片的地图图层 ID，即指定进行地图切片的图层，可以是临时图层集，也可以是当前地图中图层的组合。如果此参数缺省则对全部图层进行切片。layersID 可以是临时图层创建时templayers的ID。<br>
  *        clipRegionEnabled - {boolean} 地图显示裁剪的区域。是一个面对象，当 clipRegionEnabled = true 时有效，即地图只显示该区域覆盖的部分。<br>
@@ -43,7 +43,7 @@ export class ImageSuperMapRest extends ol.source.TileImage {
         var layerUrl = options.url + "/image.png?";
         options.serverType = options.serverType || ServerType.ISERVER;
         //为url添加安全认证信息片段
-        layerUrl = appendCredential(options.url,layerUrl, options.serverType);
+        layerUrl = appendCredential(options.url, layerUrl, options.serverType);
 
         /*
          * @function ol.source.ImageSuperMapRest.prototype.appendCredential
@@ -52,7 +52,7 @@ export class ImageSuperMapRest extends ol.source.TileImage {
          * @param serverType - {Object} 服务类型
          * @return {string} 添加生成后的新地址
          */
-        function appendCredential(id,url, serverType) {
+        function appendCredential(id, url, serverType) {
             var newUrl = url, credential, value;
             switch (serverType) {
                 case ServerType.IPORTAL:
@@ -86,8 +86,8 @@ export class ImageSuperMapRest extends ol.source.TileImage {
         }
         layerUrl += "&transparent=" + transparent;
 
-        //是否使用缓存
-        var cacheEnabled = false;
+        //是否使用缓存吗，默认为true
+        var cacheEnabled = true;
         if (options.cacheEnabled !== undefined) {
             cacheEnabled = options.cacheEnabled;
         }
@@ -97,13 +97,13 @@ export class ImageSuperMapRest extends ol.source.TileImage {
         if (options.layersID !== undefined) {
             layerUrl += "&layersID=" + options.layersID;
         }
-        //是否重定向
+        //是否重定向,默认为false
+        var redirect = false;
         if (options.redirect !== undefined) {
-            layerUrl += "&redirect=" + options.redirect;
+            redirect = options.redirect;
         }
-        if (options.cacheEnabled !== undefined) {
-            layerUrl += "&cacheEnabled=" + options.cacheEnabled;
-        }
+        layerUrl += "&redirect=" + redirect;
+
         if (options.prjCoordSys) {
             layerUrl += "prjCoordSys=" + JSON.stringify(options.prjCoordSys);
         }
@@ -142,6 +142,9 @@ export class ImageSuperMapRest extends ol.source.TileImage {
             layersID: options.layersID
         });
 
+        //存储一个cacheEnabled
+        this.cacheEnabled = cacheEnabled;
+
         if (options.tileProxy) {
             this.tileProxy = options.tileProxy;
         }
@@ -168,6 +171,10 @@ export class ImageSuperMapRest extends ol.source.TileImage {
             //支持代理
             if (me.tileProxy) {
                 url = me.tileProxy + encodeURIComponent(url);
+            }
+            //不启用缓存时启用时间戳
+            if (!me.cacheEnabled) {
+                url += "&_t=" + new Date().getTime();
             }
 
             return url;